+++ docs/release-notes.md
@@ -7,8 +7,10 @@
   - [Notes on next version](next-release-details.md)
 
 ## Version 2023.2
+- 2023.2.4 - (Sept 28, 2023)
+  - docker tag: `latest`, `2023.2`, `2023.2.4`
 - 2023.2.3 - (Sept 21, 2023)
-  - docker tag: `latest`, `2023.2`, `2023.2.3`
+  - docker tag: `2023.2.3`
 - 2023.2.2 - (Sept 19, 2023)
   - docker tag: `2023.2.2`
 - 2023.2.1 - (Sept 14, 2023)
@@ -42,6 +44,8 @@
 - [All PRs](https://github.com/caltech-ipac/firefly/pulls?q=is%3apr++milestone%3a2023.2+)
 
 ### _Patches 2023.2_
+ - 2023.2.4
+   - Fixed: Firefly support for IPAC Table's short-form data type, e.g. 'i', 'f', 'b', etc. [commit](https://github.com/Caltech-IPAC/firefly/commit/475a4aff57374bfa70b01308e63402971f1fc291)
  - 2023.2.3
    - Fixed: More Hips render issues [Firefly-1313](https://github.com/Caltech-IPAC/firefly/pull/1432)
       - missing tiles on deep zoom